deps/v8/test/inspector/debugger/wasm-stack-check-expected.txt

Tests pausing a running script and stepping

Running test: testPauseAndStep
Instantiate
Wait for script
Got wasm script: wasm://wasm/c84b7cde
Run
Expecting to pause at 61
Paused at offset 61; wasm-expression-stack: []; local: [12]
Paused at offset 62; wasm-expression-stack: []; local: [12]
Paused at offset 64; wasm-expression-stack: [12]; local: [12]
Paused at offset 66; wasm-expression-stack: [12, 1]; local: [12]
Paused at offset 67; wasm-expression-stack: [13]; local: [12]
